requirements-dev.txt
anyjson==0.3.3
asgiref==3.2.10
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-debug-toolbar==2.2
django-eav2==0.13.0
django-extensions==3.0.4
django-js-asset==1.2.2
django-mptt==0.11.0
django-mptt-admin==1.0.2
django-rosetta==0.9.5
django-tools==0.46.1
django-utils2==2.8.0
django-webpack-loader==0.7.0
easy-thumbnails==2.7
icdiff==1.9.1
idna==2.10
mysqlclient==2.0.1
packaging==20.4
Pillow==7.2.0
polib==1.1.0
pprintpp==0.4.0
progressbar2==3.51.4
pyparsing==2.4.7
python-utils==2.4.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
transliterate==1.10.2
urllib3==1.25.10
webencodings==0.5.1
XML2Dict==0.2.2
xmltodict==0.12.0
